requirements.txt
telethon>=1.43
rapidfuzz>=3.14
httpx>=0.28
python-dotenv>=1.0
# python-dotenv>=1.0
